Added Proptype and utils for testing
Added a propType error function that can be used
to fail unit tests when wanted.

Then refactored the way required props are added
to tests. This lead to the creation of a custom
matcher to test that required props are required.

Finally added on click to list of required props
for Toggle.